+++ util2/src/test/scala/org/broadinstitute/dsde/workbench/util2/WorkbenchTestSuite.scala
@@ -1,16 +1,14 @@
 package org.broadinstitute.dsde.workbench.util2
 
-import cats.effect.{ContextShift, IO, Timer}
+import cats.effect.IO
+import cats.effect.unsafe.implicits.global
 import org.scalatest.Assertion
 import org.scalatest.prop.Configuration
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 
 import scala.concurrent.Future
-import scala.concurrent.ExecutionContext.global
 
 trait WorkbenchTestSuite {
-  implicit val timer: Timer[IO] = IO.timer(global)
-  implicit val cs: ContextShift[IO] = IO.contextShift(global)
   implicit val logger = new ConsoleLogger("unit_test", LogLevel(false, false, false, true))
 
   def ioAssertion(test: => IO[Assertion]): Future[Assertion] = test.unsafeToFuture()